--- packages/design-system/src/components/Dropdown/Dropdown.types.tsx
@@ -1,14 +1,0 @@
-import type { DimensionValue } from "../../utils/type";
-
-export type DropdownOption = {
-  label: string;
-  value: string;
-};
-
-export interface Props {
-  $options: DropdownOption[];
-  value?: string;
-  $defaultValue?: string;
-  onChange?: (value: string) => void;
-  $width?: DimensionValue;
-}
